+++ rust-src/concordium_base/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Removed `try_from_cbor` and `to_cbor` from a number of types implementing CBOR serialization in favour of just 
   using `cbor::cbor_encode/decode`.
 - The serialization of `AuthorizationsV1` is fixed to be compatible with the Haskell implementation.
+- Replace concepts `TokenHolder` and `TokenGovernance` by `TokenUpdate`.
 
 ## 8.0.0-alpha (2025-06-06)